Bug #1568612 soc: sprd: debug: fix pclint warnings
[root cause  ] fix pclint warnings
[changes     ] irq_monitor
[side effects] unknown
[self test   ]: pass
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: all
[test case]: pclint pass
[reviewers   ] nianfu.bai

[change_type ] feature_bugfix
[tag_product ] common

+++ drivers/irqchip/irq-gic-v3.c
@@ -1644,7 +1644,7 @@ static ssize_t notrace hwirq_gic_write(struct file *filep,
 	else
 		enable = SPRD_NONE;
 
-	hwirq = strtoul(kbuf, NULL, 10);
+	hwirq = (unsigned int)strtoul(kbuf, NULL, 10);
 
 	if ((hwirq < 32) || (hwirq > 1019)) {
 		ret = -EINVAL;
